+++ crates/store/src/db/tests.rs
@@ -619,7 +619,9 @@ fn select_nullifiers_by_prefix() {
     const PREFIX_LEN: u32 = 16;
     let mut conn = create_db();
     // test empty table
-    let nullifiers = sql::select_nullifiers_by_prefix(&mut conn, PREFIX_LEN, &[]).unwrap();
+    let block_number0 = 0.into();
+    let nullifiers =
+        sql::select_nullifiers_by_prefix(&mut conn, PREFIX_LEN, &[], block_number0).unwrap();
     assert!(nullifiers.is_empty());
 
     // test single item
@@ -635,6 +637,7 @@ fn select_nullifiers_by_prefix() {
         &mut conn,
         PREFIX_LEN,
         &[sql::utils::get_nullifier_prefix(&nullifier1)],
+        block_number0,
     )
     .unwrap();
     assert_eq!(
@@ -662,6 +665,7 @@ fn select_nullifiers_by_prefix() {
         &mut conn,
         PREFIX_LEN,
         &[sql::utils::get_nullifier_prefix(&nullifier1)],
+        block_number0,
     )
     .unwrap();
     assert_eq!(
@@ -675,6 +679,7 @@ fn select_nullifiers_by_prefix() {
         &mut conn,
         PREFIX_LEN,
         &[sql::utils::get_nullifier_prefix(&nullifier2)],
+        block_number0,
     )
     .unwrap();
     assert_eq!(
@@ -693,6 +698,7 @@ fn select_nullifiers_by_prefix() {
             sql::utils::get_nullifier_prefix(&nullifier1),
             sql::utils::get_nullifier_prefix(&nullifier2),
         ],
+        block_number0,
     )
     .unwrap();
     assert_eq!(
@@ -714,9 +720,30 @@ fn select_nullifiers_by_prefix() {
         &mut conn,
         PREFIX_LEN,
         &[sql::utils::get_nullifier_prefix(&num_to_nullifier(3 << 48))],
+        block_number0,
     )
     .unwrap();
     assert!(nullifiers.is_empty());
+
+    // If a block number is provided, only matching nullifiers created at or after that block are
+    // returned
+    let nullifiers = sql::select_nullifiers_by_prefix(
+        &mut conn,
+        PREFIX_LEN,
+        &[
+            sql::utils::get_nullifier_prefix(&nullifier1),
+            sql::utils::get_nullifier_prefix(&nullifier2),
+        ],
+        block_number2,
+    )
+    .unwrap();
+    assert_eq!(
+        nullifiers,
+        vec![NullifierInfo {
+            nullifier: nullifier2,
+            block_num: block_number2
+        }]
+    );
 }
 
 #[test]