+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - [BREAKING] Updated minimum Rust version to 1.84.
 - [BREAKING] `Endpoint` configuration simplified to a single string (#654).
+- [BREAKING] `CheckNullifiersByPrefix` now takes a starting block number (#707).
 - [BREAKING] Changed sync state endpoint to stream the response (#685).
 
 ### Enhancements